refactor: Unpodded components. Sorted import paths (addon only).
