+++ addon-test-support/index.js
@@ -1,8 +1,8 @@
 import TablePage from './pages/ember-table';
 import registerWaiter from 'ember-raf-scheduler/test-support/register-waiter';
-import { setSetupRowCountForTest } from 'ember-table/components/ember-tbody/component';
-import { setupTHeadForTest } from 'ember-table/components/ember-thead/component';
-import { setSimpleCheckboxForTest } from 'ember-table/components/ember-td/component';
+import { setSetupRowCountForTest } from 'ember-table/components/ember-tbody';
+import { setupTHeadForTest } from 'ember-table/components/ember-thead';
+import { setSimpleCheckboxForTest } from 'ember-table/components/ember-td';
 
 function setupForTest() {
   registerWaiter();
