+++ addon/components/-private/row-wrapper.hbs
@@ -1,0 +1,1 @@
+{{yield this.api}}
