+++ addon/components/-private/scroll-indicators.hbs
