+++ addon/components/-private/scroll-indicators.js
@@ -6,8 +6,8 @@ import { bind, scheduleOnce } from '@ember/runloop';
 import { capitalize } from '@ember/string';
 import { htmlSafe } from '@ember/template';
 import { isEmpty, isNone } from '@ember/utils';
-import { addObserver } from 'ember-table/-private/utils/observer';
-import layout from './template';
+
+import { addObserver } from '../../-private/utils/observer';
 
 /**
    Computed property macro that builds the CSS styles (position, height)
@@ -122,7 +122,6 @@ const showIndicator = location => {
 };
 
 export default Component.extend({
-  layout,
   tagName: '',
 
   /**
