+++ addon/components/ember-tbody.js
@@ -1,14 +1,12 @@
 import Component from '@ember/component';
-import { scheduleOnce } from '@ember/runloop';
-import { computed } from '@ember/object';
-import { observer } from '../../-private/utils/observer';
-import { bool, readOnly, or } from '@ember/object/computed';
-
-import CollapseTree, { SELECT_MODE } from '../../-private/collapse-tree';
-import defaultTo from '../../-private/utils/default-to';
-
-import layout from './template';
 import { assert } from '@ember/debug';
+import { computed } from '@ember/object';
+import { bool, readOnly, or } from '@ember/object/computed';
+import { scheduleOnce } from '@ember/runloop';
+
+import CollapseTree, { SELECT_MODE } from '../-private/collapse-tree';
+import defaultTo from '../-private/utils/default-to';
+import { observer } from '../-private/utils/observer';
 
 let setupRowCountForTest = false;
 export function setSetupRowCountForTest(bool) {
@@ -40,7 +38,6 @@ export function setSetupRowCountForTest(bool) {
   @public
 */
 export default Component.extend({
-  layout,
   tagName: 'tbody',
 
   /**
