+++ addon/components/ember-th/resize-handle.hbs
