+++ addon/components/ember-th/sort-indicator.hbs
