+++ addon/components/ember-th/sort-indicator.js
@@ -1,6 +1,4 @@
 import Component from '@ember/component';
-import layout from './template';
-
 import { readOnly } from '@ember/object/computed';
 
 /**
@@ -26,7 +24,6 @@ import { readOnly } from '@ember/object/computed';
 */
 
 export default Component.extend({
-  layout,
   tagName: '',
 
   /**
