+++ addon/components/ember-thead.hbs
