+++ addon/components/ember-tr.js
@@ -2,10 +2,8 @@ import Component from '@ember/component';
 import { computed } from '@ember/object';
 import { readOnly } from '@ember/object/computed';
 
-import { closest } from '../../-private/utils/element';
-
-import layout from './template';
-import { SELECT_MODE } from '../../-private/collapse-tree';
+import { SELECT_MODE } from '../-private/collapse-tree';
+import { closest } from '../-private/utils/element';
 
 /**
   The table row component. This component manages row level concerns, and yields
@@ -45,7 +43,6 @@ import { SELECT_MODE } from '../../-private/collapse-tree';
   @public
 */
 export default Component.extend({
-  layout,
   tagName: 'tr',
   classNames: ['et-tr'],
   classNameBindings: ['isEven:is-even:is-odd', 'isGroupSelected', 'isSelectable', 'isSelected'],
