+++ app/components/ember-table-loading-more.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-table-loading-more/component';
+export { default } from 'ember-table/components/ember-table-loading-more';
