+++ app/components/ember-table-private/scroll-indicators.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/-private/scroll-indicators/component';
+export { default } from 'ember-table/components/-private/scroll-indicators';
