+++ app/components/ember-table.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-table/component';
+export { default } from 'ember-table/components/ember-table';
