+++ app/components/ember-td.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-td/component';
+export { default } from 'ember-table/components/ember-td';
