+++ app/components/ember-th/resize-handle.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-th/resize-handle/component';
+export { default } from 'ember-table/components/ember-th/resize-handle';
