+++ app/components/ember-th/sort-indicator.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-th/sort-indicator/component';
+export { default } from 'ember-table/components/ember-th/sort-indicator';
