+++ app/components/ember-tr.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-table/components/ember-tr/component';
+export { default } from 'ember-table/components/ember-tr';
