+++ tests/dummy/app/components/custom-row/component.js
@@ -1,4 +1,4 @@
-import EmberTableRow from 'ember-table/components/ember-tr/component';
+import EmberTableRow from 'ember-table/components/ember-tr';
 
 export default EmberTableRow.extend({
   classNames: ['custom-row'],
