+++ types/components/ember-tfoot/component.d.ts
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
 import { EmberTableColumn,  EmberTableRow, TableRowMeta } from 'ember-table';
-import EmberTrComponent from 'ember-table/components/ember-tr/component';
+import EmberTrComponent from 'ember-table/components/ember-tr';
 
 export interface EmberTfootSignature<
   RowType extends EmberTableRow,
